+++ code/tests/integrates-on-board/Core/Src/Lib/dc-motor.h
@@ -33,13 +33,13 @@ void dc_motor_backwards(dc_motor_t *motor);
 void dc_motor_reverse(dc_motor_t *motor);
 
 // set motor's speed by providing a fraction between [0, 1]
-void dc_motor_set_speed(dc_motor_t *motor, double frac);
+void dc_motor_set_speed(dc_motor_t *motor, float frac);
 
 // get the speed [0, 1] of the motor
 double dc_motor_get_speed(dc_motor_t *motor);
 
 // set a motor's speed and direction by providing a fraction [-1, 1]
-void dc_motor_set_velocity(dc_motor_t *motor, double frac);
+void dc_motor_set_velocity(dc_motor_t *motor, float frac);
 
 // get the velocity [-1, 1] of the motor
 double dc_motor_get_velocity(dc_motor_t *motor);
